# quell-backend/openapi.yaml
openapi: 3.0.3
info:
  title: Quell AI Backend (derived from routes)
  version: 1.0.0
  description: |
    OpenAPI specification for the Quell AI backend API. Endpoints cover authentication,
    store management, Shopify OAuth, chat webhook, inbox, support tickets, appearance,
    analytics, and admin metrics.
servers:
  - url: https://histioid-imogene-spinally.ngrok-free.dev
    description: Production
  - url: http://localhost:3000
    description: Local
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Store:
      type: object
      properties:
        id: { type: integer }
        shop_domain: { type: string }
        store_id: { type: string }
        widget_token: { type: string }
        product_count: { type: integer }
        status: { type: string }
        last_sync_at: { type: string, format: date-time }
        installed_at: { type: string, format: date-time }
...
security:
  - BearerAuth: []
paths:
...
  /api/stores:
    get:
      tags: [Stores]
      summary: List stores for the authenticated user
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Seller stores
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  stores:
                    type: array
                    items: { $ref: "#/components/schemas/Store" }
  /api/stores/{storeId}/sync:
    post:
      tags: [Stores]
      summary: Trigger a manual product sync for the specified store
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: storeId
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Sync initiated
  /api/stores/{storeId}:
    get:
      tags: [Stores]
      summary: Get details for a specific store
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: storeId
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Store object
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  store: { $ref: "#/components/schemas/Store" }

    delete:
      tags: [Stores]
      summary: Disconnect or deactivate a store (soft delete)
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: storeId
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Store disconnected
...
  /api/inbox:
    get:
      tags: [Inbox]
      summary: List conversation sessions (grouped) with pagination and filters
      parameters:
        - in: query
          name: store_url
          required: true
          schema: { type: string }
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: per_page
          schema: { type: integer, default: 30 }
        - in: query
          name: channel
          schema: { type: string }
        - in: query
          name: only_unread
          schema: { type: boolean }
        - in: query
          name: search
          schema: { type: string }
      responses:
        '200':
          description: Paginated session list
          content:
            application/json:
              schema:
                type: object
                properties:
                  meta:
                    type: object
                    properties:
                      total: { type: integer }
                      page: { type: integer }
                      perPage: { type: integer }
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        session_id: { type: string }
                        channel: { type: string }
                        last_message_at: { type: string }
                        last_user_message: { type: string }
                        last_bot_response: { type: string }
                        unread_count: { type: integer }
...
  /admin/stores:
    get:
      tags: [Admin]
      summary: List all stores (admin)
      parameters:
        - in: query
          name: status
          schema: { type: string }
      responses:
        '200':
          description: Stores list
...